Add InstanceRepository::delete() (#207)
