+++ app/src/Services/InstanceRepository.php
@@ -2,16 +2,19 @@
 
 namespace App\Services;
 
+use App\Model\ExcludeNotFoundDropletAction;
 use App\Model\Instance;
 use App\Model\InstanceCollection;
 use DigitalOceanV2\Api\Droplet as DropletApi;
+use DigitalOceanV2\Entity\Droplet as DropletEntity;
 use DigitalOceanV2\Exception\ExceptionInterface;
-use DigitalOceanV2\Exception\RuntimeException;
+use SmartAssert\DigitalOceanDropletConfiguration\Factory;
 
 class InstanceRepository
 {
     public function __construct(
         private DropletApi $dropletApi,
+        private Factory $dropletConfigurationFactory,
         private string $instanceCollectionTag,
         private string $instanceTag,
     ) {
@@ -20,16 +23,34 @@ class InstanceRepository
     /**
      * @throws ExceptionInterface
      */
+    public function create(): Instance
+    {
+        $configuration = $this->dropletConfigurationFactory->create();
+
+        $dropletEntity = $this->dropletApi->create(
+            $this->instanceTag,
+            $configuration->getRegion(),
+            $configuration->getSize(),
+            $configuration->getImage(),
+            $configuration->getBackups(),
+            $configuration->getIpv6(),
+            $configuration->getVpcUuid(),
+            $configuration->getSshKeys(),
+            $configuration->getUserData(),
+            $configuration->getMonitoring(),
+            $configuration->getVolumes(),
+            $configuration->getTags()
+        );
+
+        return new Instance($dropletEntity instanceof DropletEntity ? $dropletEntity : new DropletEntity());
+    }
+
+    /**
+     * @throws ExceptionInterface
+     */
     public function findAll(): InstanceCollection
     {
-        $dropletEntities = $this->dropletApi->getAll($this->instanceCollectionTag);
-
-        $instances = [];
-        foreach ($dropletEntities as $dropletEntity) {
-            $instances[] = new Instance($dropletEntity);
-        }
-
-        return new InstanceCollection($instances);
+        return $this->findWithTag($this->instanceCollectionTag);
     }
 
     /**
@@ -37,14 +58,7 @@ class InstanceRepository
      */
     public function findCurrent(): ?Instance
     {
-        $droplets = $this->dropletApi->getAll($this->instanceTag);
-        $instances = [];
-
-        foreach ($droplets as $droplet) {
-            $instances[] = new Instance($droplet);
-        }
-
-        return (new InstanceCollection($instances))->getNewest();
+        return $this->findWithTag($this->instanceTag)->getNewest();
     }
 
     /**
@@ -52,14 +66,33 @@ class InstanceRepository
      */
     public function find(int $id): ?Instance
     {
-        try {
+        return (new ExcludeNotFoundDropletAction(function (int $id) {
             return new Instance($this->dropletApi->getById($id));
-        } catch (ExceptionInterface $exception) {
-            if ($exception instanceof RuntimeException && 404 === $exception->getCode()) {
-                return null;
-            }
+        }))($id);
+    }
 
-            throw $exception;
+    /**
+     * @throws ExceptionInterface
+     */
+    public function delete(int $id): void
+    {
+        (new ExcludeNotFoundDropletAction(function (int $id) {
+            $this->dropletApi->remove($id);
+        }))($id);
+    }
+
+    /**
+     * @throws ExceptionInterface
+     */
+    private function findWithTag(string $tag): InstanceCollection
+    {
+        $dropletEntities = $this->dropletApi->getAll($tag);
+
+        $instances = [];
+        foreach ($dropletEntities as $dropletEntity) {
+            $instances[] = new Instance($dropletEntity);
         }
+
+        return new InstanceCollection($instances);
     }
 }
